requirements.txt
# Core NLP and ML libraries
transformers>=4.35.0
torch>=2.1.0
spacy>=3.7.0
nltk>=3.8.1
scikit-learn>=1.3.0

# Web framework and UI
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
jinja2>=3.1.2

# Data handling and utilities
pandas>=2.1.0
numpy>=1.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Database and storage
pyyaml>=6.0.1

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Additional NLP tools
textstat>=0.7.3
wordfreq>=3.0.3
pyenchant>=3.2.2

# Visualization
plotly>=5.17.0
